add memory_swap_max to limit the swap usage
This attempts to fix memory test failures on systems with swap enabled.
However, it won't work on systems without unified cgroup hierarchy
(for e.g. GitHub action runner), so introduce a feature flag to control it.

The new feature flag, unified_cgroup, is enabled by default.  But you may
disable it to completely disable the functions requiring a unified cgroup
hierarchy, and also disable the related tests.  Note that this is mostly
for internal testing purpose: the package builds fine with
unified_cgroup enabled on systems without it, but using the functions
requiring it on such systems will lead to runtime errors.